backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-mongoengine==1.0.0
mongoengine==0.27.0
pymongo==4.6.1
python-dotenv==1.0.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
joblib==1.4.2
gunicorn==23.0.0
